renamed sdcdevice.sdcdeviceimpl.SdcDevice to provider.providerimpml.SdcProvider
