+++ examples/ReferenceTest/reference_device.py
@@ -8,17 +8,17 @@ from decimal import Decimal
 
 from sdc11073 import wsdiscovery
 from sdc11073.location import SdcLocation
-from sdc11073.sdcdevice import SdcDevice
+from sdc11073.provider import SdcDevice
 from sdc11073.xml_types import pm_types
 from sdc11073.mdib import DeviceMdibContainer
 from sdc11073.certloader import mk_ssl_context_from_folder
 from sdc11073.xml_types.dpws_types import ThisDeviceType, ThisModelType
 from sdc11073.loghelper import LoggerAdapter
-from sdc11073.sdcdevice.components import SdcDeviceComponents
-from sdc11073.sdcdevice.subscriptionmgr_async import SubscriptionsManagerReferenceParamAsync
+from sdc11073.provider.components import SdcDeviceComponents
+from sdc11073.provider.subscriptionmgr_async import SubscriptionsManagerReferenceParamAsync
 from sdc11073.pysoap.soapclient_async import SoapClientAsync
-from sdc11073.sdcdevice.servicesfactory import DPWSHostedService
-from sdc11073.sdcdevice.servicesfactory import HostedServices, mk_dpws_hosts
+from sdc11073.provider.servicesfactory import DPWSHostedService
+from sdc11073.provider.servicesfactory import HostedServices, mk_dpws_hosts
 
 here = os.path.dirname(__file__)
 default_mdib_path = os.path.join(here, 'reference_mdib.xml')
