+++ src/sdc11073/__init__.py
@@ -3,7 +3,7 @@ from . import namespaces
 from . import mdib
 from . import location
 from . import consumer
-from . import sdcdevice
+from . import provider
 from . import commlog
 
 __version__ = '0.0.0'
