--- src/sdc11073/sdcdevice/__init__.py
@@ -1,1 +1,0 @@
-from .sdcdeviceimpl import SdcDevice
