+++ tests/test_localizationservice.py
@@ -2,7 +2,7 @@ import unittest
 from itertools import product
 
 from sdc11073.xml_types.pm_types import LocalizedText, LocalizedTextWidth
-from sdc11073.sdcdevice.porttypes import localizationservice
+from sdc11073.provider.porttypes import localizationservice
 
 
 class TestLocationService(unittest.TestCase):
